[FIX] pos_discount: move discount logic to pos_discount
Previously, the logic for retrieving discount lines was implemented in
the `point_of_sale` module, while the `discount_product_id` was managed
in the `pos_discount` module. This commit fixes the logic by moving the
_get_discount_lines method into the `pos_discount` module.

task-5875158
