+++ addons/point_of_sale/models/account_move.py
@@ -118,14 +118,3 @@ class AccountMoveLine(models.Model):
     def _compute_name(self):
         amls = self.filtered(lambda l: not l.move_id.pos_session_ids)
         super(AccountMoveLine, amls)._compute_name()
-
-    def _get_discount_lines(self):
-        lines = super()._get_discount_lines()
-        discount_line_ids = []
-        for line in self - lines:
-            pos_orders = line.move_id.sudo().pos_order_ids
-            if pos_orders and line.product_id in pos_orders.config_id.discount_product_id:
-                discount_line_ids.append(line.id)
-        if discount_line_ids:
-            lines |= self.browse(discount_line_ids)
-        return lines
