+++ addons/pos_discount/models/__init__.py
@@ -4,3 +4,4 @@
 from . import pos_config
 from . import res_config_settings
 from . import product_product
+from . import account_move_line
